feat: unify OpenClaw & Hermes into shared AgentRuntime base layer
Eliminates ~1,500 lines of copy-pasted code between the two agent-deployment
providers by extracting all shared logic into `src/Domains/Connectors/AgentRuntime/`.

New shared classes (8 files):
- `AgentRuntime\Contracts\ProviderConfig` — readonly DTO capturing all variation points
- `AgentRuntime\Enums\DeploymentStatusEnum` — single canonical enum
- `AgentRuntime\Events\AgentDeploymentStatusChanged` — single canonical broadcast event
- `AgentRuntime\Services\WorkspaceFileBuilder` — single canonical workspace markdown builder
- `AgentRuntime\SshClient` — abstract base with all SSH/SFTP/telemetry logic
- `AgentRuntime\Services\BaseDockerComposeBuilder` — abstract base for Docker config generation
- `AgentRuntime\Actions\BaseLaunchAgentOnMachineAction` — abstract base for deployment lifecycle
- `AgentRuntime\Actions\BaseDispatchAgentDeploymentAction` — abstract base for job dispatch

Each provider is now a thin subclass (~50 lines) that implements abstract methods
returning its ProviderConfig / ConfigurationEnum keys / CustomFieldEnum keys.

Deleted (6 files): per-provider DeploymentStatusEnum, WorkspaceFileBuilder,
and AgentDeploymentStatusChanged that were byte-for-byte identical.

Adding a new provider now requires only implementing abstract methods — zero
copy-paste of deployment logic.
