+++ src/Domains/Connectors/AgentRuntime/Enums/DeploymentStatusEnum.php
@@ -2,8 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Connectors\Hermes\Enums;
+namespace Kanvas\Connectors\AgentRuntime\Enums;
 
+/**
+ * Canonical deployment lifecycle states shared by all agent runtime providers.
+ *
+ * Replaces the per-provider copies (OpenClaw\Enums\DeploymentStatusEnum,
+ * Hermes\Enums\DeploymentStatusEnum) which were byte-for-byte identical.
+ */
 enum DeploymentStatusEnum: string
 {
     case PENDING = 'pending';
