+++ src/Domains/Connectors/AgentRuntime/Events/AgentDeploymentStatusChanged.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Connectors\OpenClaw\Events;
+namespace Kanvas\Connectors\AgentRuntime\Events;
 
 use Illuminate\Broadcasting\Channel;
 use Illuminate\Broadcasting\InteractsWithSockets;
@@ -12,6 +12,12 @@ use Illuminate\Queue\SerializesModels;
 use Kanvas\Intelligence\Agents\Models\AgentDeployment;
 use Override;
 
+/**
+ * Broadcast event emitted whenever an AgentDeployment transitions to a new status.
+ *
+ * Replaces the per-provider copies (OpenClaw\Events\AgentDeploymentStatusChanged,
+ * Hermes\Events\AgentDeploymentStatusChanged) which were byte-for-byte identical.
+ */
 class AgentDeploymentStatusChanged implements ShouldBroadcastNow
 {
     use Dispatchable;
@@ -45,11 +51,11 @@ class AgentDeploymentStatusChanged implements ShouldBroadcastNow
     public function broadcastWith(): array
     {
         return [
-            'deployment_id' => $this->deployment->id,
-            'agent_id' => $this->deployment->agent_id,
-            'status' => $this->deployment->status,
+            'deployment_id'   => $this->deployment->id,
+            'agent_id'        => $this->deployment->agent_id,
+            'status'          => $this->deployment->status,
             'previous_status' => $this->previousStatus,
-            'error_message' => $this->deployment->error_message,
+            'error_message'   => $this->deployment->error_message,
         ];
     }
 }
