+++ src/Domains/Connectors/Hermes/Jobs/LaunchAgentJob.php
@@ -12,7 +12,7 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Kanvas\Connectors\Hermes\Actions\LaunchAgentOnMachineAction;
-use Kanvas\Connectors\Hermes\Events\AgentDeploymentStatusChanged;
+use Kanvas\Connectors\AgentRuntime\Events\AgentDeploymentStatusChanged;
 use Kanvas\Intelligence\Agents\Models\Agent;
 use Kanvas\Intelligence\Agents\Models\AgentDeployment;
 use Kanvas\Intelligence\Agents\Models\AgentMachine;
