+++ src/Domains/Connectors/OpenClaw/Actions/DeployAgentAction.php
@@ -7,7 +7,7 @@ namespace Kanvas\Connectors\OpenClaw\Actions;
 use Baka\Contracts\AppInterface;
 use Baka\Contracts\CompanyInterface;
 use Kanvas\Connectors\OpenClaw\Enums\CustomFieldEnum;
-use Kanvas\Connectors\OpenClaw\Services\WorkspaceFileBuilder;
+use Kanvas\Connectors\AgentRuntime\Services\WorkspaceFileBuilder;
 use Kanvas\Connectors\OpenClaw\SshClient;
 use Kanvas\Intelligence\Agents\Models\Agent;
 use Throwable;
