+++ src/Domains/Connectors/OpenClaw/Actions/RebuildSharedImageAction.php
@@ -24,22 +24,23 @@ class RebuildSharedImageAction
 
     public function execute(): void
     {
-        $client = SshClient::fromMachine($this->machine);
-        $imageName = DockerComposeBuilder::getSharedImageName($this->app);
-        $imageDir = DockerComposeBuilder::getSharedImageDir($this->app);
+        $builder   = new DockerComposeBuilder();
+        $client    = SshClient::fromMachine($this->machine);
+        $imageName = $builder->getSharedImageName($this->app);
+        $imageDir  = $builder->getSharedImageDir($this->app);
 
         try {
             $client->exec('sudo mkdir -p ' . escapeshellarg($imageDir));
 
             $client->writeFileAsUser(
                 $imageDir . '/Dockerfile',
-                DockerComposeBuilder::buildDockerfile($this->app),
+                $builder->buildDockerfile($this->app),
                 'root',
             );
 
             $client->writeFileAsUser(
                 $imageDir . '/entrypoint.sh',
-                DockerComposeBuilder::buildEntrypoint(),
+                $builder->buildEntrypoint(),
                 'root',
             );
             $client->exec('sudo chmod +x ' . escapeshellarg($imageDir . '/entrypoint.sh'));
