+++ src/Domains/Connectors/OpenClaw/Actions/SyncAgentWorkspaceAction.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Kanvas\Connectors\OpenClaw\Actions;
 
-use Kanvas\Connectors\OpenClaw\Services\WorkspaceFileBuilder;
+use Kanvas\Connectors\AgentRuntime\Services\WorkspaceFileBuilder;
 use Kanvas\Connectors\OpenClaw\SshClient;
 use Kanvas\Intelligence\Agents\Models\Agent;
 use Kanvas\Intelligence\Agents\Models\AgentDeployment;
