--- src/Domains/Connectors/OpenClaw/Enums/DeploymentStatusEnum.php
@@ -1,16 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Kanvas\Connectors\OpenClaw\Enums;
-
-enum DeploymentStatusEnum: string
-{
-    case PENDING = 'pending';
-    case PROVISIONING = 'provisioning';
-    case RUNNING = 'running';
-    case UPDATING = 'updating';
-    case STOPPED = 'stopped';
-    case FAILED = 'failed';
-    case TERMINATED = 'terminated';
-}
