+++ src/Domains/Connectors/OpenClaw/Jobs/UpdateOpenClawForUserJob.php
@@ -9,8 +9,8 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
-use Kanvas\Connectors\OpenClaw\Enums\DeploymentStatusEnum;
-use Kanvas\Connectors\OpenClaw\Events\AgentDeploymentStatusChanged;
+use Kanvas\Connectors\AgentRuntime\Enums\DeploymentStatusEnum;
+use Kanvas\Connectors\AgentRuntime\Events\AgentDeploymentStatusChanged;
 use Kanvas\Connectors\OpenClaw\SshClient;
 use Kanvas\Exceptions\ValidationException;
 use Kanvas\Intelligence\Agents\Models\AgentDeployment;
